@akien-mga 2025-04-01
Note that there's more code in Color (like `Color::html`) which could possibly benefit from similar optimizations around the handling of `#`.